requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appnope==0.1.3
APScheduler==3.10.1
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.14.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
blobfile==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cryptography==39.0.2
curlify==2.2.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-tinymce==3.6.1
docopt==0.6.2
-e git+https://github.com/openai/evals.git@19bfdba1fd9027f6818672e35615f45584de8b02#egg=evals
exceptiongroup==1.1.0
executing==1.2.0
facebook-business==16.0.2
facebook-sdk==3.1.0
filelock==3.9.0
fire==0.5.0
Flask==2.2.2
Flask-HTTPAuth==4.7.0
fonttools==4.39.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json2xml==3.21.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langdetect==1.0.9
lazy-object-proxy==1.9.0
lxml==4.9.2
lz4==4.3.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mock==5.0.1
mock-open==1.4.0
multidict==6.0.4
mypy==1.1.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pick==2.2.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.2.18
pipreqs==0.4.11
platformdirs==2.6.2
portalocker==2.7.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml-env==1.2.1
pyarrow==10.0.1
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.6
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.16.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
PySocks==1.7.1
python-barcode==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
pyzstd==0.15.4
qrcode==7.4.2
regex==2022.10.31
requests==2.28.2
sacrebleu==2.3.1
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.0.1
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
termcolor==2.2.0
tiktoken==0.3.1
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.13
virtualenv==20.20.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yarg==0.1.9
yarl==1.8.2
